# api/bff-openapi.yaml
...
openapi: "3.0.3"
info:
  title: Skateboard UI-backend API
  version: "1.0.0"
...
paths:
...
  /api/home/videos:
    get:
      tags: [home]
      operationId: getHomeVideos
      summary: >-
        Get the videos eligible for the Home dashboard (unshuffled — the
        client randomizes for display; see README_HOME_DASHBOARD.md §4/§23)
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_HOME
      responses:
        '200':
          description: Home dashboard videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HomeVideoResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_HOME required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/home/featured-player:
    get:
      tags: [home-featured-player]
      operationId: getHomeFeaturedPlayer
      summary: >-
        Get the effective Featured Player content for the Home dashboard, or
        204 when none is active/configured/resolvable
      security:
        - bearerAuth: []
      x-required-permissions:
        - FUNC_TAB_HOME
      responses:
        '200':
          description: Featured Player content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeFeaturedPlayerResponse'
        '204':
          description: No Featured Player to show
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden – FUNC_TAB_HOME required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FeaturedContentSource:
      type: string
      description: >-
        Logical content source owning the referenced contentId. Only PODCAST
        exists today; new sources are added here as they're onboarded.
      enum: [PODCAST]
...
    HomePlayerPosition:
      type: string
      enum: [TOP, BOTTOM]
...
    HomeFeaturedPlayerPlayback:
      type: object
      properties:
        type:
          type: string
          description: Playback adapter the mini player should use, e.g. SPOTIFY_EMBED or YOUTUBE.
        reference:
          type: string
          description: Provider-specific playback reference (e.g. the Spotify/YouTube URL) — opaque to the frontend beyond picking the adapter for `type`.
    HomeFeaturedPlayerResponse:
      type: object
      properties:
        id:
          type: string
        source:
          $ref: '#/components/schemas/FeaturedContentSource'
        title:
          type: string
        subtitle:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        playback:
          $ref: '#/components/schemas/HomeFeaturedPlayerPlayback'
        position:
          $ref: '#/components/schemas/HomePlayerPosition'
    HomeVideoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          description: Required for navigation — Video Details (/podcast/{slug}) has no by-id route.
        title:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        thumbnailWidth:
          type: integer
          nullable: true
          description: >
            Intrinsic pixel width of thumbnailUrl, captured by podcast-be's
            YouTube sync. Lets the Home masonry gallery size a tile from its real
            aspect ratio without measuring the image first. Null on videos whose
            dimensions were never captured — clients must fall back to probing.
        thumbnailHeight:
          type: integer
          nullable: true
          description: Intrinsic pixel height of thumbnailUrl; see thumbnailWidth.
        youtubeVideoId:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
          description: Category slug the video was fetched under (null in ALL mode).
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time